Refactored after sonarLint

+++ src/main/java/inventory/controller/MainScreenController.java
@@ -3,7 +3,6 @@ package inventory.controller;
 
 import inventory.model.Part;
 import inventory.model.Product;
-import inventory.service.InventoryService;
 import inventory.service.PartService;
 import inventory.service.ProductService;
 import javafx.event.ActionEvent;
@@ -35,15 +34,13 @@ public class MainScreenController implements Initializable,Controller {
     private static int modifyProductIndex;
     
     // Declare methods
-    public static int getModifyPartIndex() {
+    static int getModifyPartIndex() {
         return modifyPartIndex;
     }
     
-    public static int getModifyProductIndex() {
+    static int getModifyProductIndex() {
         return modifyProductIndex;
     }
-
-    //private InventoryService service;
 
     private PartService partService;
     private ProductService productService;
@@ -88,7 +85,6 @@ public class MainScreenController implements Initializable,Controller {
     public MainScreenController(){}
 
     public void setServices(PartService partService, ProductService productService){
-       // this.service=service;
         this.partService = partService;
         this.productService = productService;
         partsTableView.setItems(partService.getAllParts());
@@ -123,7 +119,6 @@ public class MainScreenController implements Initializable,Controller {
     private void displayScene(ActionEvent event, String source) throws IOException {
         stage = (Stage)((Button)event.getSource()).getScene().getWindow();
         FXMLLoader loader= new FXMLLoader(getClass().getResource(source));
-        //scene = FXMLLoader.load(getClass().getResource(source));
         scene = loader.load();
         Controller ctrl=loader.getController();
         ctrl.setServices(partService,productService);
@@ -145,12 +140,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Part Deletion?");
         alert.setContentText("Are you sure you want to delete part " + part.getName() + " from parts?");
         Optional<ButtonType> result = alert.showAndWait();
-
-        if (result.get() == ButtonType.OK) {
-            System.out.println("Part deleted.");
-            partService.deletePart(part);
-        } else {
-            System.out.println("Canceled part deletion.");
+        if(result.isPresent()) {
+            if (result.get() == ButtonType.OK) {
+                System.out.println("Part deleted.");
+                partService.deletePart(part);
+            } else {
+                System.out.println("Canceled part deletion.");
+            }
         }
     }
 
@@ -168,12 +164,14 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Product Deletion?");
         alert.setContentText("Are you sure you want to delete product " + product.getName() + " from products?");
         Optional<ButtonType> result = alert.showAndWait();
-        
-        if (result.get() == ButtonType.OK) {
-            productService.deleteProduct(product);
-            System.out.println("Product " + product.getName() + " was removed.");
-        } else {
-            System.out.println("Product " + product.getName() + " was not removed.");
+
+        if(result.isPresent()) {
+            if (result.get() == ButtonType.OK) {
+                productService.deleteProduct(product);
+                System.out.println("Product " + product.getName() + " was removed.");
+            } else {
+                System.out.println("Product " + product.getName() + " was not removed.");
+            }
         }
     }
 
@@ -236,11 +234,13 @@ public class MainScreenController implements Initializable,Controller {
         alert.setHeaderText("Confirm Exit");
         alert.setContentText("Are you sure you want to exit?");
         Optional<ButtonType> result = alert.showAndWait();
-        if(result.get() == ButtonType.OK) {
-            System.out.println("Ok selected. Program exited");
-            System.exit(0);
-        } else {
-            System.out.println("Cancel clicked.");
+        if(result.isPresent()) {
+            if (result.get() == ButtonType.OK) {
+                System.out.println("Ok selected. Program exited");
+                System.exit(0);
+            } else {
+                System.out.println("Cancel clicked.");
+            }
         }
     }
